Added a simple test for premium calculation and implemented the service contract

+++ src/PremiumCalculator.Api.Data/Entities/Rating.cs
@@ -6,7 +6,7 @@ namespace PremiumCalculator.Api.Data.Entities
     public class Rating : BaseEntity
     {
         public string Label { get; set; }
-        public float Factor { get; set; }
+        public double Factor { get; set; }
         public virtual List<Occupation> Occupations { get; set; }
     }
 }
